+++ src/main/java/calculator/operators/Plus.java
@@ -48,6 +48,6 @@ public final class Plus extends Operation
         return new MyNotANumber();
     if (l instanceof ComplexNumber || r instanceof ComplexNumber)
         return new ComplexNumber(l.getReal() + r.getReal(), l.getImaginary() + r.getImaginary());
-      return new MyNumber((float) l.getValue() + r.getValue());
+      return new MyNumber(l.getValue() + r.getValue());
   }
 }
